Refactor : entities api를 READ만 남기고, 나머지는 model로 분리
