+++ src/entities/comments/index.ts
@@ -1,2 +1,4 @@
 export * from "./api/index"
 export * from "./api/types"
+export * from "./model/index"
+export * from "./model/types"
