+++ src/entities/posts/api/types.ts
@@ -23,35 +23,6 @@ export interface PostsData {
   limit: number
 }
 
-export interface CreatePostsRequest {
-  title: string
-  body: string
-  userId: number
-}
-
-export interface CreatePostsResponse {
-  id: number
-  title: string
-  body: string
-  userId: number
-}
-
-export interface UpdatePostsRequest {
-  id: number
-}
-
-export interface UpdatePostsResponse {
-  id: number
-  title: string
-  body: string
-  userId: number
-  tags: Array<string>
-  reactions: {
-    likes: number
-    dislikes: number
-  }
-}
-
 export interface PostsTags {
   slug: string
   name: string
