+++ src/entities/posts/model/types.ts
@@ -1,0 +1,28 @@
+export interface CreatePostsRequest {
+  title: string
+  body: string
+  userId: number
+}
+
+export interface CreatePostsResponse {
+  id: number
+  title: string
+  body: string
+  userId: number
+}
+
+export interface UpdatePostsRequest {
+  id: number
+}
+
+export interface UpdatePostsResponse {
+  id: number
+  title: string
+  body: string
+  userId: number
+  tags: Array<string>
+  reactions: {
+    likes: number
+    dislikes: number
+  }
+}
